feat: New videos App
This app will be dedicated to reinforce some of the details already
studied in the course. It will be a simple version of a website like
Youtube, where a user can search a term and the website
should return a list of videos and when clicking on a video, it will be
reproduced.