feat: Header
The Header component will use the Link component to display
the different options in the Menu.
The Link component will render the name and will handle the click event which triggers the path change.

The Rute component will be lisening to this event and save the change in the state. This will make the component to re-render